refactor: delete BlockFaceBehaviour
