--- Assets/Scripts/BlockFaceBehaviour.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 26d3224b9def85647b13ad6b57e05c75
-timeCreated: 1519040476
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
